refactor: divide domain
