+++ src/collections/entities/collection.entity.ts
@@ -4,7 +4,7 @@ import { CoreEntity } from '../../common/entities/core.entity';
 import { NestedContent } from './nested-content.entity';
 import { User } from '../../users/entities/user.entity';
 import { Column, Entity, ManyToOne, OneToMany, RelationId } from 'typeorm';
-import { Category } from '../../contents/entities/category.entity';
+import { Category } from '../../categories/category.entity';
 
 @Entity()
 export class Collection extends CoreEntity {
