+++ src/contents/contents.module.ts
@@ -1,35 +1,26 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import {
-  CategoryController,
-  ContentsController,
-  TestController,
-} from './contents.controller';
-import { CategoryService, ContentsService } from './contents.service';
-import { Category } from './entities/category.entity';
+import { ContentsController } from './contents.controller';
+import { ContentsService } from './contents.service';
+import { Category } from '../categories/category.entity';
 import { Content } from './entities/content.entity';
 import { CategoryUtil } from './util/category.util';
 import { ContentRepository } from './repository/content.repository';
-import { CategoryRepository } from './repository/category.repository';
+import { CategoryRepository } from '../categories/category.repository';
 import { UsersModule } from '../users/users.module';
 import { ContentUtil } from './util/content.util';
 import { OpenaiModule } from '../openai/openai.module';
 
 @Module({
-  imports: [
-    TypeOrmModule.forFeature([Content, Category]),
-    UsersModule,
-    OpenaiModule,
-  ],
-  controllers: [ContentsController, CategoryController, TestController],
+  imports: [TypeOrmModule.forFeature([Content]), UsersModule, OpenaiModule],
+  controllers: [ContentsController],
   providers: [
     ContentsService,
-    CategoryService,
     ContentRepository,
+    ContentUtil,
     CategoryRepository,
     CategoryUtil,
-    ContentUtil,
   ],
-  exports: [ContentsService, CategoryRepository],
+  exports: [ContentsService],
 })
 export class ContentsModule {}
